Built-in

Below is the list of tranforms that are are available in this package.

Select

TableTransforms.Select — Type

Select(col₁, col₂, ..., colₙ)
Select([col₁, col₂, ..., colₙ])
Select((col₁, col₂, ..., colₙ))

The transform that selects columns col₁, col₂, ..., colₙ.

Select(col₁ => newcol₁, col₂ => newcol₂, ..., colₙ => newcolₙ)

Selects the columns col₁, col₂, ..., colₙ and rename them to newcol₁, newcol₂, ..., newcolₙ.

Select(regex)

Selects the columns that match with regex.

Examples

Select(1, 3, 5)
Select([:a, :c, :e])
Select(("a", "c", "e"))
Select(1 => :x, 3 => :y)
Select(:a => :x, :b => :y)
Select("a" => "x", "b" => "y")
Select(r"[ace]")

Reject

TableTransforms.Reject — Type

Reject(col₁, col₂, ..., colₙ)
Reject([col₁, col₂, ..., colₙ])
Reject((col₁, col₂, ..., colₙ))

The transform that discards columns col₁, col₂, ..., colₙ.

Reject(regex)

Discards the columns that match with regex.

Examples

Reject(:b, :d, :f)
Reject(["b", "d", "f"])
Reject((2, 4, 6))
Reject(r"[bdf]")

Rename

TableTransforms.Rename — Type

Rename(:col₁ => :newcol₁, :col₂ => :newcol₂, ..., :colₙ => :newcolₙ)

The transform that renames col₁, col₂, ..., colₙ to newcol₁, newcol₂, ..., newcolₙ.

Examples

Rename(1 => :x, 3 => :y)
Rename(:a => :x, :c => :y)
Rename("a" => "x", "c" => "y")

StdNames

TableTransforms.StdNames — Type

StdNames(spec)

Standardizes column names according to given spec. Default to :upper case specification.

Specs

:upper - Uppercase, e.g. COLUMNNAME

:camel - Camelcase, e.g. ColumnName

:snake - Snakecase, e.g. column_name

Sort

TableTransforms.Sort — Type

Sort(col₁, col₂, ..., colₙ; kwargs...)
Sort([col₁, col₂, ..., colₙ]; kwargs...)
Sort((col₁, col₂, ..., colₙ); kwargs...)

Sort the rows of selected columns col₁, col₂, ..., colₙ by forwarding the kwargs to the sortperm function.

Sort(regex; kwargs...)

Sort the rows of columns that match with regex.

Examples

Sort(:a)
Sort(:a, :c, rev=true)
Sort([1, 3, 5], by=row -> abs.(row))
Sort(("a", "c", "e"))
Sort(r"[ace]")

Sample

TableTransforms.Sample — Type

Sample(size, [weights]; replace=true, ordered=false, rng=GLOBAL_RNG)

Sample size rows of table using weights with or without replacement depending on the option replace. The option ordered can be used to return samples in the same order of the original table.

Examples

Sample(1_000)
Sample(1_000, replace=false)
Sample(1_000, replace=false, ordered=true)

# with rng
using Random
rng = MersenneTwister(2)
Sample(1_000, rng=rng)

# with weights
Sample(10, rand(100))

Filter

TableTransforms.Filter — Type

Filter(func)

Filters the table returning only the rows where the func returns true.

Examples

Filter(row -> sum(row) > 10)
Filter(row -> row.a == true && row.b < 30)

Notes

The schema of the table is preserved by the transform.

DropMissing

TableTransforms.DropMissing — Type

DropMissing()
DropMissing(:)

Drop all rows with missing values in table.

DropMissing(col₁, col₂, ..., colₙ)
DropMissing([col₁, col₂, ..., colₙ])
DropMissing((col₁, col₂, ..., colₙ))

Drop all rows with missing values in selected columns col₁, col₂, ..., colₙ.

DropMissing(regex)

Drop all rows with missing values in columns that match with regex.

Examples

DropMissing()
DropMissing("b", "c", "e")
DropMissing([2, 3, 5])
DropMissing((:b, :c, :e))
DropMissing(r"[bce]")

Notes

The transform can alter the element type of columns from Union{Missing,T} to T.

Replace

TableTransforms.Replace — Type

Replace(old₁ => new₁, old₂ => new₂, ..., oldₙ => newₙ)

Replaces oldᵢ value with newᵢ value in the table.

Examples

Replace(1 => -1, 5 => -5)
Replace(1 => 1.5, 5 => 5.5, 4 => true)

Coalesce

TableTransforms.Coalesce — Type

Coalesce(; value)

Replaces all missing values from the table with value.

Coalesce(col₁, col₂, ..., colₙ; value)
Coalesce([col₁, col₂, ..., colₙ]; value)
Coalesce((col₁, col₂, ..., colₙ); value)

Replaces all missing values from the columns col₁, col₂, ..., colₙ with value.

Coalesce(regex; value)

Replaces all missing values from the columns that match with regex with value.

Examples

Coalesce(value=0)
Coalesce(1, 3, 5, value=1)
Coalesce([:a, :c, :e], value=2)
Coalesce(("a", "c", "e"), value=3)
Coalesce(r"[ace]", value=4)

Notes

The transform can alter the element type of columns from Union{Missing,T} to T.

Coerce

TableTransforms.Coerce — Type

Coerce(pairs, tight=false, verbosity=1)

Return a copy of the table, ensuring that the scientific types of the columns match the new specification.

This transform wraps the ScientificTypes.coerce function. Please see their docstring for more details.

Examples

using ScientificTypes
Coerce(:col1 => Continuous, :col2 => Count)

Levels

TableTransforms.Levels — Type

Levels(col₁ => levels₁, col₂ => levels₂, ..., colₙ => levelsₙ; ordered=nothing)

Convert columns col₁, col₂, ..., colₙ to categorical arrays with given levels levels₁, levels₂, ..., levelsₙ. Optionally, specify which columns are ordered.

Examples

Levels(1 => 1:3, 2 => ["a", "b"], ordered=r"a")
Levels(:a => 1:3, :b => ["a", "b"], ordered=[:a])
Levels("a" => 1:3, "b" => ["a", "b"], ordered=["b"])

OneHot

TableTransforms.OneHot — Type

OneHot(col)

Transforms categorical column col into one-hot columns of levels returned by the levels function of CategoricalArrays.jl.

Examples

OneHot(1)
OneHot(:a)
OneHot("a")

Identity

TransformsBase.Identity — Type

Identity()

The identity transform that maps any object to itself.

Center

TableTransforms.Center — Type

Center()

Applies the center transform to all columns of the table. The center transform of the column x, with mean μ, is defined by x .- μ.

Center(col₁, col₂, ..., colₙ)
Center([col₁, col₂, ..., colₙ])
Center((col₁, col₂, ..., colₙ))

Applies the Center transform on columns col₁, col₂, ..., colₙ.

Center(regex)

Applies the Center transform on columns that match with regex.

Examples

Center(1, 3, 5)
Center([:a, :c, :e])
Center(("a", "c", "e"))
Center(r"[ace]")

Scale

TableTransforms.Scale — Type

Scale(; low=0.25, high=0.75)

Applies the Scale transform to all columns of the table. The scale transform of the column x is defined by (x .- xl) ./ (xh - xl), where xl = quantile(x, low) and xh = quantile(x, high).

Scale(col₁, col₂, ..., colₙ; low=0.25, high=0.75)
Scale([col₁, col₂, ..., colₙ]; low=0.25, high=0.75)
Scale((col₁, col₂, ..., colₙ); low=0.25, high=0.75)

Applies the Scale transform on columns col₁, col₂, ..., colₙ.

Scale(regex; low=0.25, high=0.75)

Applies the Scale transform on columns that match with regex.

Examples

Scale()
Scale(low=0, high=1)
Scale(low=0.3, high=0.7)
Scale(1, 3, 5, low=0, high=1)
Scale([:a, :c, :e], low=0.3, high=0.7)
Scale(("a", "c", "e"), low=0.25, high=0.75)
Scale(r"[ace]", low=0.3, high=0.7)

Notes

The low and high values are restricted to the interval [0, 1].

MinMax

TableTransforms.MinMax — Function

MinMax()

Applies the MinMax transform to all columns of the table. The MinMax transform is equivalent to Scale(low=0, high=1).

MinMax(col₁, col₂, ..., colₙ)
MinMax([col₁, col₂, ..., colₙ])
MinMax((col₁, col₂, ..., colₙ))

Applies the MinMax transform on columns col₁, col₂, ..., colₙ.

MinMax(regex)

Applies the MinMax transform on columns that match with regex.

Examples

MinMax(1, 3, 5)
MinMax([:a, :c, :e])
MinMax(("a", "c", "e"))
MinMax(r"[ace]")

See also Scale.

Interquartile

TableTransforms.Interquartile — Function

Interquartile()

Applies the Interquartile transform to all columns of the table. The Interquartile transform is equivalent to Scale(low=0.25, high=0.75).

Interquartile(col₁, col₂, ..., colₙ)
Interquartile([col₁, col₂, ..., colₙ])
Interquartile((col₁, col₂, ..., colₙ))

Applies the Interquartile transform on columns col₁, col₂, ..., colₙ.

Interquartile(regex)

Applies the Interquartile transform on columns that match with regex.

Examples

Interquartile(1, 3, 5)
Interquartile([:a, :c, :e])
Interquartile(("a", "c", "e"))
Interquartile(r"[ace]")

See also Scale.

ZScore

TableTransforms.ZScore — Type

ZScore()

Applies the z-score transform (a.k.a. normal score) to all columns of the table. The z-score transform of the column x, with mean μ and standard deviation σ, is defined by (x .- μ) ./ σ.

ZScore(col₁, col₂, ..., colₙ)
ZScore([col₁, col₂, ..., colₙ])
ZScore((col₁, col₂, ..., colₙ))

Applies the ZScore transform on columns col₁, col₂, ..., colₙ.

ZScore(regex)

Applies the ZScore transform on columns that match with regex.

Examples

ZScore(1, 3, 5)
ZScore([:a, :c, :e])
ZScore(("a", "c", "e"))
ZScore(r"[ace]")

Quantile

TableTransforms.Quantile — Type

Quantile(; dist=Normal())

The quantile transform to a given distribution.

Quantile(col₁, col₂, ..., colₙ; dist=Normal())
Quantile([col₁, col₂, ..., colₙ]; dist=Normal())
Quantile((col₁, col₂, ..., colₙ); dist=Normal())

Applies the Quantile transform on columns col₁, col₂, ..., colₙ.

Quantile(regex; dist=Normal())

Applies the Quantile transform on columns that match with regex.

Examples

using Distributions

Quantile()
Quantile(dist=Normal())
Quantile(1, 3, 5, dist=Beta())
Quantile([:a, :c, :e], dist=Gamma())
Quantile(("a", "c", "e"), dist=Beta())
Quantile(r"[ace]", dist=Normal())

Functional

TableTransforms.Functional — Type

Functional(func)

The transform that applies a func elementwise.

Functional(col₁ => func₁, col₂ => func₂, ..., colₙ => funcₙ)

Apply the corresponding funcᵢ function to each colᵢ column.

Examples

Functional(cos)
Functional(sin)
Functional(1 => cos, 2 => sin)
Functional(:a => cos, :b => sin)
Functional("a" => cos, "b" => sin)

EigenAnalysis

TableTransforms.EigenAnalysis — Type

EigenAnalysis(proj; maxdim=nothing, pratio=1.0)

The eigenanalysis of the covariance with a given projection proj. Optionally specify the maximum number of dimensions in the output maxdim and the percentage of variance to retain pratio. Default to all dimensions of the input.

Projections

:V - Uncorrelated variables (PCA transform)

:VD - Uncorrelated variables and variance one (DRS transform)

:VDV - Uncorrelated variables and variance one (SDS transformation)

The :V projection used in the PCA transform projects the data on the eigenvectors V of the covariance matrix.

The :VD projection used in the DRS transform. Similar to the :V projection, but the eigenvectors are multiplied by the squared inverse of the eigenvalues D.

The :VDV projection used in the SDS transform. Similar to the :VD transform, but the data is projected back to the basis of the original variables using the Vᵀ matrix.

See https://geostatisticslessons.com/lessons/sphereingmaf for more details about these three variants of eigenanalysis.

Examples

EigenAnalysis(:V)
EigenAnalysis(:VD)
EigenAnalysis(:VDV)
EigenAnalysis(:V, maxdim=3)
EigenAnalysis(:VD, pratio=0.99)
EigenAnalysis(:VDV, maxdim=3, pratio=0.99)

PCA

TableTransforms.PCA — Function

PCA(; maxdim=nothing, pratio=1.0)

The PCA transform is a shortcut for ZScore() → EigenAnalysis(:V; maxdim, pratio).

See also: ZScore, EigenAnalysis.

Examples

PCA(maxdim=2)
PCA(pratio=0.86)
PCA(maxdim=2, pratio=0.86)

DRS

TableTransforms.DRS — Function

DRS(; maxdim=nothing, pratio=1.0)

The DRS transform is a shortcut for ZScore() → EigenAnalysis(:VD; maxdim, pratio).

See also: ZScore, EigenAnalysis.

Examples

DRS(maxdim=3)
DRS(pratio=0.87)
DRS(maxdim=3, pratio=0.87)

SDS

TableTransforms.SDS — Function

SDS(; maxdim=nothing, pratio=1.0)

The SDS transform is a shortcut for ZScore() → EigenAnalysis(:VDV; maxdim, pratio).

See also: ZScore, EigenAnalysis.

Examples

SDS()
SDS(maxdim=4)
SDS(pratio=0.88)
SDS(maxdim=4, pratio=0.88)

RowTable

TableTransforms.RowTable — Type

RowTable()

The transform that applies the function Tables.rowtable to to the input table.

ColTable

TableTransforms.ColTable — Type

ColTable()

The transform that applies the function Tables.columntable to to the input table.