Built-in

Below is the list of tranforms that are are available in this package.

Select

TableTransforms.Select — Type

Select(col₁, col₂, ..., colₙ)
Select([col₁, col₂, ..., colₙ])
Select((col₁, col₂, ..., colₙ))

The transform that selects columns col₁, col₂, ..., colₙ.

Select(regex)

Selects the columns that match with regex.

Examples

Select(1, 3, 5)
Select([:a, :c, :e])
Select(("a", "c", "e"))
Select(r"[ace]")

Reject

TableTransforms.Reject — Type

Reject(col₁, col₂, ..., colₙ)
Reject([col₁, col₂, ..., colₙ])
Reject((col₁, col₂, ..., colₙ))

The transform that discards columns col₁, col₂, ..., colₙ.

Reject(regex)

Discards the columns that match with regex.

Examples

Reject(:b, :d, :f)
Reject(["b", "d", "f"])
Reject((2, 4, 6))
Reject(r"[bdf]")

Filter

TableTransforms.Filter — Type

Filter(func)

Filters the table returning only the rows where the func returns true.

Examples

Filter(row -> sum(row) > 10)
Filter(row -> row.a == true && row.b < 30)

Notes

The schema of the table is preserved by the transform.

DropMissing

TableTransforms.DropMissing — Type

DropMissing()
DropMissing(:)

Drop all rows with missing values in table.

DropMissing(col₁, col₂, ..., colₙ)
DropMissing([col₁, col₂, ..., colₙ])
DropMissing((col₁, col₂, ..., colₙ))

Drop all rows with missing values in selected columns col₁, col₂, ..., colₙ.

DropMissing(regex)

Drop all rows with missing values in columns that match with regex.

Examples

DropMissing()
DropMissing("b", "c", "e")
DropMissing([2, 3, 5])
DropMissing((:b, :c, :e))
DropMissing(r"[bce]")

Notes

The transform can alter the element type of columns from Union{Missing,T} to T.

Rename

TableTransforms.Rename — Type

Rename(:col₁ => :newcol₁, :col₂ => :newcol₂, ..., :col₁ => :newcolₙ)

The transform that renames col₁ to newcol₁, col₂ to newcol₂, ...

Examples

Rename(:a => :x, :c => :y)
Rename("a" => "x", "c" => "y")

Replace

TableTransforms.Replace — Type

Replace(old₁ => new₁, old₂ => new₂, ..., oldₙ => newₙ)

Replaces oldᵢ value with newᵢ value in the table.

Examples

Replace(1 => -1, 5 => -5)
Replace(1 => 1.5, 5 => 5.5, 4 => true)

Coalesce

TableTransforms.Coalesce — Type

Coalesce(value)

Replaces all missing values from the table with value.

Examples

Coalesce(0)

Notes

The transform can alter the element type of columns from Union{Missing,T} to T.

Coerce

TableTransforms.Coerce — Type

Coerce(pairs, tight=false, verbosity=1)

Return a copy of the table, ensuring that the scientific types of the columns match the new specification.

This transform wraps the ScientificTypes.coerce function. Please see their docstring for more details.

Examples

Coerce(:col1 => Continuous, :col2 => Count)

RowTable

TableTransforms.RowTable — Type

RowTable()

The transform that applies the function Tables.rowtable to to the input table.

ColTable

TableTransforms.ColTable — Type

ColTable()

The transform that applies the function Tables.columntable to to the input table.

Identity

TableTransforms.Identity — Type

Identity()

The identity transform x -> x.

Center

TableTransforms.Center — Type

Center()

Applies the center transform to all columns of the table. The center transform of the column x, with mean μ, is defined by x .- μ.

Scale

TableTransforms.Scale — Type

Scale(; low=0.25, high=0.75)

Applies the scale transform to all columns of the table. The scale transform of the column x is defined by (x .- xl) ./ (xh - xl), where xl = quantile(x, low) and xh = quantile(x, high).

Examples

Scale()
Scale(low=0, high=1)
Scale(low=0.3, high=0.7)

Notes

The low and high values are restricted to the interval [0, 1].

MinMax

TableTransforms.MinMax — Function

MinMax()

The transform that is equivalent to Scale(low=0, high=1).

See also Scale.

Interquartile

TableTransforms.Interquartile — Function

Interquartile()

The transform that is equivalent to Scale(low=0.25, high=0.75).

See also Scale.

ZScore

TableTransforms.ZScore — Type

ZScore()

Applies the z-score transform (a.k.a. normal score) to all columns of the table. The z-score transform of the column x, with mean μ and standard deviation σ, is defined by (x .- μ) ./ σ.

Quantile

TableTransforms.Quantile — Type

Quantile(distribution=Normal())

The quantile transform to a given distribution.

Examples

Quantile()
Quantile(distribution=Normal())

Functional

TableTransforms.Functional — Type

Functional(func)

The transform that applies a func elementwise.

Functional(col₁ => func₁, col₂ => func₂, ..., colₙ => funcₙ)

Apply the corresponding funcᵢ function to each colᵢ column.

Examples

Functional(cos)
Functional(sin)
Functional(:a => cos, :b => sin)
Functional("a" => cos, "b" => sin)

EigenAnalysis

TableTransforms.EigenAnalysis — Type

EigenAnalysis(proj)

The eigenanalysis of the covariance with a given projection proj.

Projections

:V - Uncorrelated variables (PCA transform)

:VD - Uncorrelated variables and variance one (DRS transform)

:VDV - Uncorrelated variables and variance one (SDS transformation)

The :V projection used in the PCA transform projects the data on the eigenvectors V of the covariance matrix.

The :VD projection used in the DRS transform. Similar to the :V projection, but the eigenvectors are multiplied by the squared inverse of the eigenvalues D.

The :VDV projection used in the SDS transform. Similar to the :VD transform, but the data is projected back to the basis of the original variables using the Vᵀ matrix.

See https://geostatisticslessons.com/lessons/sphereingmaf for more details about these three variants of eigenanalysis.

Examples

EigenAnalysis(:V)
EigenAnalysis(:VD)
EigenAnalysis(:VDV)

PCA

TableTransforms.PCA — Function

PCA()

The PCA transform is a shortcut for ZScore() → EigenAnalysis(:V).

See also: ZScore, EigenAnalysis.

DRS

TableTransforms.DRS — Function

DRS()

The DRS transform is a shortcut for ZScore() → EigenAnalysis(:VD).

See also: ZScore, EigenAnalysis.

SDS

TableTransforms.SDS — Function

SDS()

The SDS transform is a shortcut for ZScore() → EigenAnalysis(:VDV).

See also: ZScore, EigenAnalysis.

Sequential

TableTransforms.Sequential — Type

Sequential(transforms)

A transform where transforms are applied in sequence.

Examples

Scale(low=0.2, high=0.8) → EigenAnalysis(:VDV)
Select(:x, :z) → ZScore() → EigenAnalysis(:V) → Scale(low=0, high=1)

Parallel

TableTransforms.Parallel — Type

Parallel(transforms)

A transform where transforms are applied in parallel.

Examples

Scale(low=0.3, high=0.6) ⊔ EigenAnalysis(:VDV)
ZScore() ⊔ EigenAnalysis(:V)